move fixtures around
